backend/requirements.txt
asgiref==3.9.1
black==25.1.0
certifi==2025.8.3
click==8.2.1
cloudinary==1.44.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
gunicorn==23.0.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
PyJWT==2.10.1
python-dotenv==1.1.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
